Crate google-calendar3

This documentation was generated from calendar crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

acl

delete, get, insert, list, patch, update and watch

calendar list

delete, get, insert, list, patch, update and watch

calendars

clear, delete, get, insert, patch and update

channels

stop

colors

get

events

delete, get, import, insert, instances, list, move, patch, quick add, update and watch

freebusy

query

settings

get, list and watch

Subscription supported by ...

list settings

list events

list calendar list

watch events

instances events

watch settings

watch acl

list acl

watch calendar list

Everything else about the calendar v3 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.events().delete(...).doit()
let r = hub.events().insert(...).doit()
let r = hub.events().instances(...).doit()
let r = hub.events().quick_add(...).doit()
let r = hub.events().patch(...).doit()
let r = hub.events().import(...).doit()
let r = hub.events().move_(...).doit()
let r = hub.events().update(...).doit()
let r = hub.events().watch(...).doit()
let r = hub.events().get(...).doit()
let r = hub.events().list(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-calendar3 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-calendar3" as calendar3;
use calendar3::Channel;
use calendar3::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use calendar3::CalendarHub;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CalendarHub::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Channel = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.events().watch(&req, "calendarId")
             .updated_min("ea")
             .time_zone("no")
             .time_min("justo")
             .time_max("justo")
             .sync_token("et")
             .single_events(true)
             .show_hidden_invitations(true)
             .show_deleted(false)
             .add_shared_extended_property("Lorem")
             .q("et")
             .add_private_extended_property("duo")
             .page_token("aliquyam")
             .order_by("sea")
             .max_results(-55)
             .max_attendees(-75)
             .i_cal_uid("erat")
             .always_include_email(false)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Acl
There is no detailed description.

AclDeleteCallBuilder
Deletes an access control rule.

AclGetCallBuilder
Returns an access control rule.

AclInsertCallBuilder
Creates an access control rule.

AclListCallBuilder
Returns the rules in the access control list for the calendar.

AclMethodsBuilder
A builder providing access to all methods supported on acl resources. It is not used directly, but through the CalendarHub hub.

AclPatchCallBuilder
Updates an access control rule. This method supports patch semantics.

AclRule
There is no detailed description.

AclRuleScope
The scope of the rule.

AclUpdateCallBuilder
Updates an access control rule.

AclWatchCallBuilder
Watch for changes to ACL resources.

Calendar
There is no detailed description.

CalendarClearCallBuilder
Clears a primary calendar. This operation deletes all events associated with the primary calendar of an account.

CalendarDeleteCallBuilder
Deletes a secondary calendar. Use calendars.clear for clearing all events on primary calendars.

CalendarGetCallBuilder
Returns metadata for a calendar.

CalendarHub
Central instance to access all CalendarHub related resource activities

CalendarInsertCallBuilder
Creates a secondary calendar.

CalendarList
There is no detailed description.

CalendarListDeleteCallBuilder
Deletes an entry on the user's calendar list.

CalendarListEntry
There is no detailed description.

CalendarListEntryNotificationSettings
The notifications that the authenticated user is receiving for this calendar.

CalendarListGetCallBuilder
Returns an entry on the user's calendar list.

CalendarListInsertCallBuilder
Adds an entry to the user's calendar list.

CalendarListListCallBuilder
Returns entries on the user's calendar list.

CalendarListMethodsBuilder
A builder providing access to all methods supported on calendarList resources. It is not used directly, but through the CalendarHub hub.

CalendarListPatchCallBuilder
Updates an entry on the user's calendar list. This method supports patch semantics.

CalendarListUpdateCallBuilder
Updates an entry on the user's calendar list.

CalendarListWatchCallBuilder
Watch for changes to CalendarList resources.

CalendarMethodsBuilder
A builder providing access to all methods supported on calendar resources. It is not used directly, but through the CalendarHub hub.

CalendarNotification
There is no detailed description.

CalendarPatchCallBuilder
Updates metadata for a calendar. This method supports patch semantics.

CalendarUpdateCallBuilder
Updates metadata for a calendar.

Channel
There is no detailed description.

ChannelMethodsBuilder
A builder providing access to all methods supported on channel resources. It is not used directly, but through the CalendarHub hub.

ChannelStopCallBuilder
Stop watching resources through this channel

ColorDefinition
There is no detailed description.

ColorGetCallBuilder
Returns the color definitions for calendars and events.

ColorMethodsBuilder
A builder providing access to all methods supported on color resources. It is not used directly, but through the CalendarHub hub.

Colors
There is no detailed description.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Error
There is no detailed description.

Event
There is no detailed description.

EventAttachment
There is no detailed description.

EventAttendee
There is no detailed description.

EventCreator
The creator of the event. Read-only.

EventDateTime
There is no detailed description.

EventDeleteCallBuilder
Deletes an event.

EventExtendedProperties
Extended properties of the event.

EventGadget
A gadget that extends this event.

EventGetCallBuilder
Returns an event.

EventImportCallBuilder
Imports an event. This operation is used to add a private copy of an existing event to a calendar.

EventInsertCallBuilder
Creates an event.

EventInstanceCallBuilder
Returns instances of the specified recurring event.

EventListCallBuilder
Returns events on the specified calendar.

EventMethodsBuilder
A builder providing access to all methods supported on event resources. It is not used directly, but through the CalendarHub hub.

EventMoveCallBuilder
Moves an event to another calendar, i.e. changes an event's organizer.

EventOrganizer
The organizer of the event. If the organizer is also an attendee, this is indicated with a separate entry in attendees with the organizer field set to True. To change the organizer, use the move operation. Read-only, except when importing an event.

EventPatchCallBuilder
Updates an event. This method supports patch semantics.

EventQuickAddCallBuilder
Creates an event based on a simple text string.

EventReminder
There is no detailed description.

EventReminders
Information about the event's reminders for the authenticated user.

EventSource
Source of an event from which it was created; for example a web page, an email message or any document identifiable by an URL using HTTP/HTTPS protocol. Accessible only by the creator of the event.

EventUpdateCallBuilder
Updates an event.

EventWatchCallBuilder
Watch for changes to Events resources.

Events
There is no detailed description.

FreeBusyCalendar
There is no detailed description.

FreeBusyGroup
There is no detailed description.

FreeBusyRequest
There is no detailed description.

FreeBusyRequestItem
There is no detailed description.

FreeBusyResponse
There is no detailed description.

FreebusyMethodsBuilder
A builder providing access to all methods supported on freebusy resources. It is not used directly, but through the CalendarHub hub.

FreebusyQueryCallBuilder
Returns free/busy information for a set of calendars.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Setting
There is no detailed description.

SettingGetCallBuilder
Returns a single user setting.

SettingListCallBuilder
Returns all user settings for the authenticated user.

SettingMethodsBuilder
A builder providing access to all methods supported on setting resources. It is not used directly, but through the CalendarHub hub.

SettingWatchCallBuilder
Watch for changes to Settings resources.

Settings
There is no detailed description.

TimePeriod
There is no detailed description.

Enums

Result
A universal result type used as return for all action method results.

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

ResponseResult
Identifies types which are used in API responses.

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.