Module google_calendar3::api

Structs

Acl
There is no detailed description.

AclDeleteCall
Deletes an access control rule.

AclGetCall
Returns an access control rule.

AclInsertCall
Creates an access control rule.

AclListCall
Returns the rules in the access control list for the calendar.

AclMethods
A builder providing access to all methods supported on acl resources. It is not used directly, but through the CalendarHub hub.

AclPatchCall
Updates an access control rule. This method supports patch semantics.

AclRule
There is no detailed description.

AclRuleScope
The extent to which calendar access is granted by this ACL rule.

AclUpdateCall
Updates an access control rule.

AclWatchCall
Watch for changes to ACL resources.

Calendar
There is no detailed description.

CalendarClearCall
Clears a primary calendar. This operation deletes all events associated with the primary calendar of an account.

CalendarDeleteCall
Deletes a secondary calendar. Use calendars.clear for clearing all events on primary calendars.

CalendarGetCall
Returns metadata for a calendar.

CalendarHub
Central instance to access all CalendarHub related resource activities

CalendarInsertCall
Creates a secondary calendar.

CalendarList
There is no detailed description.

CalendarListDeleteCall
Removes a calendar from the user's calendar list.

CalendarListEntry
There is no detailed description.

CalendarListEntryNotificationSettings
The notifications that the authenticated user is receiving for this calendar.

CalendarListGetCall
Returns a calendar from the user's calendar list.

CalendarListInsertCall
Inserts an existing calendar into the user's calendar list.

CalendarListListCall
Returns the calendars on the user's calendar list.

CalendarListMethods
A builder providing access to all methods supported on calendarList resources. It is not used directly, but through the CalendarHub hub.

CalendarListPatchCall
Updates an existing calendar on the user's calendar list. This method supports patch semantics.

CalendarListUpdateCall
Updates an existing calendar on the user's calendar list.

CalendarListWatchCall
Watch for changes to CalendarList resources.

CalendarMethods
A builder providing access to all methods supported on calendar resources. It is not used directly, but through the CalendarHub hub.

CalendarNotification
There is no detailed description.

CalendarPatchCall
Updates metadata for a calendar. This method supports patch semantics.

CalendarUpdateCall
Updates metadata for a calendar.

Channel
There is no detailed description.

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the CalendarHub hub.

ChannelStopCall
Stop watching resources through this channel

ColorDefinition
There is no detailed description.

ColorGetCall
Returns the color definitions for calendars and events.

ColorMethods
A builder providing access to all methods supported on color resources. It is not used directly, but through the CalendarHub hub.

Colors
There is no detailed description.

ConferenceData
There is no detailed description.

ConferenceParameters
There is no detailed description.

ConferenceParametersAddOnParameters
There is no detailed description.

ConferenceProperties
There is no detailed description.

ConferenceRequestStatus
There is no detailed description.

ConferenceSolution
There is no detailed description.

ConferenceSolutionKey
There is no detailed description.

CreateConferenceRequest
There is no detailed description.

EntryPoint
There is no detailed description.

Error
There is no detailed description.

Event
There is no detailed description.

EventAttachment
There is no detailed description.

EventAttendee
There is no detailed description.

EventCreator
The creator of the event. Read-only.

EventDateTime
There is no detailed description.

EventDeleteCall
Deletes an event.

EventExtendedProperties
Extended properties of the event.

EventGadget
A gadget that extends this event. Gadgets are deprecated; this structure is instead only used for returning birthday calendar metadata.

EventGetCall
Returns an event.

EventImportCall
Imports an event. This operation is used to add a private copy of an existing event to a calendar.

EventInsertCall
Creates an event.

EventInstanceCall
Returns instances of the specified recurring event.

EventListCall
Returns events on the specified calendar.

EventMethods
A builder providing access to all methods supported on event resources. It is not used directly, but through the CalendarHub hub.

EventMoveCall
Moves an event to another calendar, i.e. changes an event's organizer.

EventOrganizer
The organizer of the event. If the organizer is also an attendee, this is indicated with a separate entry in attendees with the organizer field set to True. To change the organizer, use the move operation. Read-only, except when importing an event.

EventPatchCall
Updates an event. This method supports patch semantics.

EventQuickAddCall
Creates an event based on a simple text string.

EventReminder
There is no detailed description.

EventReminders
Information about the event's reminders for the authenticated user.

EventSource
Source from which the event was created. For example, a web page, an email message or any document identifiable by an URL with HTTP or HTTPS scheme. Can only be seen or modified by the creator of the event.

EventUpdateCall
Updates an event.

EventWatchCall
Watch for changes to Events resources.

Events
There is no detailed description.

FreeBusyCalendar
There is no detailed description.

FreeBusyGroup
There is no detailed description.

FreeBusyRequest
There is no detailed description.

FreeBusyRequestItem
There is no detailed description.

FreeBusyResponse
There is no detailed description.

FreebusyMethods
A builder providing access to all methods supported on freebusy resources. It is not used directly, but through the CalendarHub hub.

FreebusyQueryCall
Returns free/busy information for a set of calendars.

Setting
There is no detailed description.

SettingGetCall
Returns a single user setting.

SettingListCall
Returns all user settings for the authenticated user.

SettingMethods
A builder providing access to all methods supported on setting resources. It is not used directly, but through the CalendarHub hub.

SettingWatchCall
Watch for changes to Settings resources.

Settings
There is no detailed description.

TimePeriod
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.