Struct google_calendar3::api::EventListCall

pub struct EventListCall<'a, C> where
    C: 'a,  { /* fields omitted */ }

Returns events on the specified calendar.

A builder for the list method supported by a event resource. It is not used directly, but through a EventMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.events().list("calendarId")
             .updated_min("accusam")
             .time_zone("amet")
             .time_min("erat")
             .time_max("dolores")
             .sync_token("erat")
             .single_events(false)
             .show_hidden_invitations(true)
             .show_deleted(true)
             .add_shared_extended_property("et")
             .q("At")
             .add_private_extended_property("dolor")
             .page_token("et")
             .order_by("sit")
             .max_results(-81)
             .max_attendees(-10)
             .i_cal_uid("nonumy")
             .always_include_email(true)
             .doit().await;

Implementations

impl<'a, C> EventListCall<'a, C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub async fn doit(self) -> Result<(Response<Body>, Events)>

Perform the operation you have build so far.

pub fn calendar_id(self, new_value: &str) -> EventListCall<'a, C>

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.

Sets the calendar id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn updated_min(self, new_value: &str) -> EventListCall<'a, C>

Lower bound for an event's last modification time (as a RFC3339 timestamp) to filter by. When specified, entries deleted since this time will always be included regardless of showDeleted. Optional. The default is not to filter by last modification time.

Sets the updated min query property to the given value.

pub fn time_zone(self, new_value: &str) -> EventListCall<'a, C>

Time zone used in the response. Optional. The default is the time zone of the calendar.

Sets the time zone query property to the given value.

pub fn time_min(self, new_value: &str) -> EventListCall<'a, C>

Lower bound (exclusive) for an event's end time to filter by. Optional. The default is not to filter by end time. Must be an RFC3339 timestamp with mandatory time zone offset, for example, 2011-06-03T10:00:00-07:00, 2011-06-03T10:00:00Z. Milliseconds may be provided but are ignored. If timeMax is set, timeMin must be smaller than timeMax.

Sets the time min query property to the given value.

pub fn time_max(self, new_value: &str) -> EventListCall<'a, C>

Upper bound (exclusive) for an event's start time to filter by. Optional. The default is not to filter by start time. Must be an RFC3339 timestamp with mandatory time zone offset, for example, 2011-06-03T10:00:00-07:00, 2011-06-03T10:00:00Z. Milliseconds may be provided but are ignored. If timeMin is set, timeMax must be greater than timeMin.

Sets the time max query property to the given value.

pub fn sync_token(self, new_value: &str) -> EventListCall<'a, C>

Token obtained from the nextSyncToken field returned on the last page of results from the previous list request. It makes the result of this list request contain only entries that have changed since then. All events deleted since the previous list request will always be in the result set and it is not allowed to set showDeleted to False. There are several query parameters that cannot be specified together with nextSyncToken to ensure consistency of the client state.

These are:

iCalUID

orderBy

privateExtendedProperty

q

sharedExtendedProperty

timeMin

timeMax

updatedMin If the syncToken expires, the server will respond with a 410 GONE response code and the client should clear its storage and perform a full synchronization without any syncToken. Learn more about incremental synchronization. Optional. The default is to return all entries.

Sets the sync token query property to the given value.

pub fn single_events(self, new_value: bool) -> EventListCall<'a, C>

Whether to expand recurring events into instances and only return single one-off events and instances of recurring events, but not the underlying recurring events themselves. Optional. The default is False.

Sets the single events query property to the given value.

pub fn show_hidden_invitations(self, new_value: bool) -> EventListCall<'a, C>

Whether to include hidden invitations in the result. Optional. The default is False.

Sets the show hidden invitations query property to the given value.

pub fn show_deleted(self, new_value: bool) -> EventListCall<'a, C>

Whether to include deleted events (with status equals "cancelled") in the result. Cancelled instances of recurring events (but not the underlying recurring event) will still be included if showDeleted and singleEvents are both False. If showDeleted and singleEvents are both True, only single instances of deleted events (but not the underlying recurring events) are returned. Optional. The default is False.

Sets the show deleted query property to the given value.

pub fn add_shared_extended_property(
self,
new_value: &str
) -> EventListCall<'a, C>

Extended properties constraint specified as propertyName=value. Matches only shared properties. This parameter might be repeated multiple times to return events that match all given constraints.

Append the given value to the shared extended property query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

pub fn q(self, new_value: &str) -> EventListCall<'a, C>

Free text search terms to find events that match these terms in any field, except for extended properties. Optional.

Sets the q query property to the given value.

pub fn add_private_extended_property(
self,
new_value: &str
) -> EventListCall<'a, C>

Extended properties constraint specified as propertyName=value. Matches only private properties. This parameter might be repeated multiple times to return events that match all given constraints.

Append the given value to the private extended property query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

pub fn page_token(self, new_value: &str) -> EventListCall<'a, C>

Token specifying which result page to return. Optional.

Sets the page token query property to the given value.

pub fn order_by(self, new_value: &str) -> EventListCall<'a, C>

The order of the events returned in the result. Optional. The default is an unspecified, stable order.

Sets the order by query property to the given value.

pub fn max_results(self, new_value: i32) -> EventListCall<'a, C>

Maximum number of events returned on one result page. The number of events in the resulting page may be less than this value, or none at all, even if there are more events matching the query. Incomplete pages can be detected by a non-empty nextPageToken field in the response. By default the value is 250 events. The page size can never be larger than 2500 events. Optional.

Sets the max results query property to the given value.

pub fn max_attendees(self, new_value: i32) -> EventListCall<'a, C>

The maximum number of attendees to include in the response. If there are more than the specified number of attendees, only the participant is returned. Optional.

Sets the max attendees query property to the given value.

pub fn i_cal_uid(self, new_value: &str) -> EventListCall<'a, C>

Specifies event ID in the iCalendar format to be included in the response. Optional.

Sets the i cal uid query property to the given value.

pub fn always_include_email(self, new_value: bool) -> EventListCall<'a, C>

Deprecated and ignored. A value will always be returned in the email field for the organizer, creator and attendees, even if no real email address is available (i.e. a generated, non-working value will be provided).

Sets the always include email query property to the given value.

pub fn delegate(self, new_value: &'a mut dyn Delegate) -> EventListCall<'a, C>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> EventListCall<'a, C> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

alt (query-string) - Data format for the response.

fields (query-string) - Selector specifying which fields to include in a partial response.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

oauth_token (query-string) - OAuth 2.0 token for the current user.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

quotaUser (query-string) - An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

userIp (query-string) - Deprecated. Please use quotaUser instead.

pub fn add_scope<T, S>(self, scope: T) -> EventListCall<'a, C> where
T: Into<Option<S>>,
S: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::EventReadonly.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope. If None is specified, then all scopes will be removed and no default scope will be used either. In that case, you have to specify your API-key using the key parameter (see the param() function for details).

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.