Struct google-calendar3::Calendar

pub struct Calendar {
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub description: Option<String>,
    pub id: Option<String>,
    pub time_zone: Option<String>,
    pub location: Option<String>,
    pub summary: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

clear calendars (none)

get calendars (response)

update calendars (request|response)

patch calendars (request|response)

insert calendars (request|response)

delete calendars (none)

Fields

kind

Type of the resource ("calendar#calendar").

etag

ETag of the resource.

description

Description of the calendar. Optional.

id

Identifier of the calendar.

time_zone

The time zone of the calendar. (Formatted as an IANA Time Zone Database name, e.g. "Europe/Zurich".) Optional.

location

Geographic location of the calendar as free-form text. Optional.

summary

Title of the calendar.