Struct google_calendar3::CalendarListMethods

pub struct CalendarListMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on calendarList resources. It is not used directly, but through the CalendarHub hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_calendar3 as calendar3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use calendar3::CalendarHub;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CalendarHub::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)`, `update(...)` and `watch(...)`
// to build up your call.
let rb = hub.calendar_list();

Methods

impl<'a, C, A> CalendarListMethods<'a, C, A>

fn update(&self, request: CalendarListEntry, calendar_id: &str) -> CalendarListUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an entry on the user's calendar list.

Arguments

request - No description provided.

calendarId - Calendar identifier.

fn delete(&self, calendar_id: &str) -> CalendarListDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an entry on the user's calendar list.

Arguments

calendarId - Calendar identifier.

fn get(&self, calendar_id: &str) -> CalendarListGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns an entry on the user's calendar list.

Arguments

calendarId - Calendar identifier.

fn list(&self) -> CalendarListListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns entries on the user's calendar list.

fn insert(&self, request: CalendarListEntry) -> CalendarListInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Adds an entry to the user's calendar list.

Arguments

request - No description provided.

fn patch(&self, request: CalendarListEntry, calendar_id: &str) -> CalendarListPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an entry on the user's calendar list. This method supports patch semantics.

Arguments

request - No description provided.

calendarId - Calendar identifier.

fn watch(&self, request: Channel) -> CalendarListWatchCall<'a, C, A>

Create a builder to help you perform the following task:

Watch for changes to CalendarList resources.

Arguments

request - No description provided.