Struct google_calendar3::CalendarMethods

pub struct CalendarMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on calendar resources. It is not used directly, but through the CalendarHub hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_calendar3 as calendar3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use calendar3::CalendarHub;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CalendarHub::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `clear(...)`, `delete(...)`, `get(...)`, `insert(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.calendars();

Methods

impl<'a, C, A> CalendarMethods<'a, C, A>

fn patch(&self, request: Calendar, calendar_id: &str) -> CalendarPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates metadata for a calendar. This method supports patch semantics.

Arguments

request - No description provided.

calendarId - Calendar identifier.

fn delete(&self, calendar_id: &str) -> CalendarDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a secondary calendar. Use calendars.clear for clearing all events on primary calendars.

Arguments

calendarId - Calendar identifier.

fn get(&self, calendar_id: &str) -> CalendarGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns metadata for a calendar.

Arguments

calendarId - Calendar identifier.

fn clear(&self, calendar_id: &str) -> CalendarClearCall<'a, C, A>

Create a builder to help you perform the following task:

Clears a primary calendar. This operation deletes all events associated with the primary calendar of an account.

Arguments

calendarId - Calendar identifier.

fn insert(&self, request: Calendar) -> CalendarInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a secondary calendar.

Arguments

request - No description provided.

fn update(&self, request: Calendar, calendar_id: &str) -> CalendarUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates metadata for a calendar.

Arguments

request - No description provided.

calendarId - Calendar identifier.