Struct google_calendar3::CalendarNotification

pub struct CalendarNotification {
    pub type_: Option<String>,
    pub method: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
type_: Option<String>

The type of notification. Possible values are: - "eventCreation" - Notification sent when a new event is put on the calendar. - "eventChange" - Notification sent when an event is changed. - "eventCancellation" - Notification sent when an event is cancelled. - "eventResponse" - Notification sent when an event is changed. - "agenda" - An agenda with the events of the day (sent out in the morning).

method: Option<String>

The method used to deliver the notification. Possible values are: - "email" - Reminders are sent via email. - "sms" - Reminders are sent via SMS. This value is read-only and is ignored on inserts and updates. SMS reminders are only available for Google Apps for Work, Education, and Government customers.