Struct google_calendar3::Channel

pub struct Channel {
    pub resource_uri: Option<String>,
    pub kind: Option<String>,
    pub resource_id: Option<String>,
    pub id: Option<String>,
    pub token: Option<String>,
    pub params: Option<HashMap<String, String>>,
    pub expiration: Option<String>,
    pub address: Option<String>,
    pub type_: Option<String>,
    pub payload: Option<bool>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

stop channels (request)

watch settings (request|response)

watch acl (request|response)

watch calendar list (request|response)

watch events (request|response)

Fields
resource_uri: Option<String>

A version-specific identifier for the watched resource.

kind: Option<String>

Identifies this as a notification channel used to watch for changes to a resource. Value: the fixed string "api#channel".

resource_id: Option<String>

An opaque ID that identifies the resource being watched on this channel. Stable across different API versions.

id: Option<String>

A UUID or similar unique string that identifies this channel.

token: Option<String>

An arbitrary string delivered to the target address with each notification delivered over this channel. Optional.

params: Option<HashMap<String, String>>

Additional parameters controlling delivery channel behavior. Optional.

expiration: Option<String>

Date and time of notification channel expiration, expressed as a Unix timestamp, in milliseconds. Optional.

address: Option<String>

The address where notifications are delivered for this channel.

type_: Option<String>

The type of delivery mechanism used for this channel.

payload: Option<bool>

A Boolean value to indicate whether payload is wanted. Optional.