Struct google_calendar3::EventAttachment

pub struct EventAttachment {
    pub mime_type: Option<String>,
    pub icon_link: Option<String>,
    pub file_url: Option<String>,
    pub file_id: Option<String>,
    pub title: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

mime_type

Internet media type (MIME type) of the attachment.

icon_link

URL link to the attachment's icon. Read-only.

file_url

URL link to the attachment. For adding Google Drive file attachments use the same format as in alternateLink property of the Files resource in the Drive API.

file_id

ID of the attached file. Read-only. For Google Drive files, this is the ID of the corresponding Files resource entry in the Drive API.

title

Attachment title.