Struct google_calendar3::EventWatchCall

pub struct EventWatchCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Watch for changes to Events resources.

A builder for the watch method supported by a event resource. It is not used directly, but through a EventMethods instance.

Example

Instantiate a resource method builder

use calendar3::Channel;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Channel::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.events().watch(req, "calendarId")
             .updated_min("Lorem")
             .time_zone("clita")
             .time_min("invidunt")
             .time_max("eirmod")
             .sync_token("At")
             .single_events(false)
             .show_hidden_invitations(true)
             .show_deleted(false)
             .add_shared_extended_property("sit")
             .q("takimata")
             .add_private_extended_property("elitr")
             .page_token("nonumy")
             .order_by("rebum.")
             .max_results(-6)
             .max_attendees(-50)
             .i_cal_uid("diam")
             .always_include_email(true)
             .doit();

Methods

impl<'a, C, A> EventWatchCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, Channel)>

Perform the operation you have build so far.

fn request(self, new_value: Channel) -> EventWatchCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn calendar_id(self, new_value: &str) -> EventWatchCall<'a, C, A>

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.

Sets the calendar id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn updated_min(self, new_value: &str) -> EventWatchCall<'a, C, A>

Lower bound for an event's last modification time (as a RFC3339 timestamp) to filter by. When specified, entries deleted since this time will always be included regardless of showDeleted. Optional. The default is not to filter by last modification time.

Sets the updated min query property to the given value.

fn time_zone(self, new_value: &str) -> EventWatchCall<'a, C, A>

Time zone used in the response. Optional. The default is the time zone of the calendar.

Sets the time zone query property to the given value.

fn time_min(self, new_value: &str) -> EventWatchCall<'a, C, A>

Lower bound (inclusive) for an event's end time to filter by. Optional. The default is not to filter by end time. Must be an RFC3339 timestamp with mandatory time zone offset, e.g., 2011-06-03T10:00:00-07:00, 2011-06-03T10:00:00Z. Milliseconds may be provided but will be ignored.

Sets the time min query property to the given value.

fn time_max(self, new_value: &str) -> EventWatchCall<'a, C, A>

Upper bound (exclusive) for an event's start time to filter by. Optional. The default is not to filter by start time. Must be an RFC3339 timestamp with mandatory time zone offset, e.g., 2011-06-03T10:00:00-07:00, 2011-06-03T10:00:00Z. Milliseconds may be provided but will be ignored.

Sets the time max query property to the given value.

fn sync_token(self, new_value: &str) -> EventWatchCall<'a, C, A>

Token obtained from the nextSyncToken field returned on the last page of results from the previous list request. It makes the result of this list request contain only entries that have changed since then. All events deleted since the previous list request will always be in the result set and it is not allowed to set showDeleted to False. There are several query parameters that cannot be specified together with nextSyncToken to ensure consistency of the client state.

These are: - iCalUID - orderBy - privateExtendedProperty - q - sharedExtendedProperty - timeMin - timeMax - updatedMin If the syncToken expires, the server will respond with a 410 GONE response code and the client should clear its storage and perform a full synchronization without any syncToken. Learn more about incremental synchronization. Optional. The default is to return all entries.

Sets the sync token query property to the given value.

fn single_events(self, new_value: bool) -> EventWatchCall<'a, C, A>

Whether to expand recurring events into instances and only return single one-off events and instances of recurring events, but not the underlying recurring events themselves. Optional. The default is False.

Sets the single events query property to the given value.

fn show_hidden_invitations(self, new_value: bool) -> EventWatchCall<'a, C, A>

Whether to include hidden invitations in the result. Optional. The default is False.

Sets the show hidden invitations query property to the given value.

fn show_deleted(self, new_value: bool) -> EventWatchCall<'a, C, A>

Whether to include deleted events (with status equals "cancelled") in the result. Cancelled instances of recurring events (but not the underlying recurring event) will still be included if showDeleted and singleEvents are both False. If showDeleted and singleEvents are both True, only single instances of deleted events (but not the underlying recurring events) are returned. Optional. The default is False.

Sets the show deleted query property to the given value.

fn add_shared_extended_property(self, new_value: &str) -> EventWatchCall<'a, C, A>

Extended properties constraint specified as propertyName=value. Matches only shared properties. This parameter might be repeated multiple times to return events that match all given constraints.

Append the given value to the shared extended property query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn q(self, new_value: &str) -> EventWatchCall<'a, C, A>

Free text search terms to find events that match these terms in any field, except for extended properties. Optional.

Sets the q query property to the given value.

fn add_private_extended_property(self, new_value: &str) -> EventWatchCall<'a, C, A>

Extended properties constraint specified as propertyName=value. Matches only private properties. This parameter might be repeated multiple times to return events that match all given constraints.

Append the given value to the private extended property query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn page_token(self, new_value: &str) -> EventWatchCall<'a, C, A>

Token specifying which result page to return. Optional.

Sets the page token query property to the given value.

fn order_by(self, new_value: &str) -> EventWatchCall<'a, C, A>

The order of the events returned in the result. Optional. The default is an unspecified, stable order.

Sets the order by query property to the given value.

fn max_results(self, new_value: i32) -> EventWatchCall<'a, C, A>

Maximum number of events returned on one result page. By default the value is 250 events. The page size can never be larger than 2500 events. Optional.

Sets the max results query property to the given value.

fn max_attendees(self, new_value: i32) -> EventWatchCall<'a, C, A>

The maximum number of attendees to include in the response. If there are more than the specified number of attendees, only the participant is returned. Optional.

Sets the max attendees query property to the given value.

fn i_cal_uid(self, new_value: &str) -> EventWatchCall<'a, C, A>

Specifies event ID in the iCalendar format to be included in the response. Optional.

Sets the i cal uid query property to the given value.

fn always_include_email(self, new_value: bool) -> EventWatchCall<'a, C, A>

Whether to always include a value in the email field for the organizer, creator and attendees, even if no real email is available (i.e. a generated, non-working value will be provided). The use of this option is discouraged and should only be used by clients which cannot handle the absence of an email address value in the mentioned places. Optional. The default is False.

Sets the always include email query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> EventWatchCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> EventWatchCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> EventWatchCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.