Struct google_calendar3::Events

pub struct Events {
    pub next_page_token: Option<String>,
    pub kind: Option<String>,
    pub default_reminders: Option<Vec<EventReminder>>,
    pub description: Option<String>,
    pub items: Option<Vec<Event>>,
    pub updated: Option<String>,
    pub summary: Option<String>,
    pub etag: Option<String>,
    pub next_sync_token: Option<String>,
    pub time_zone: Option<String>,
    pub access_role: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

instances events (response)

list events (response)

Fields

next_page_token

Token used to access the next page of this result. Omitted if no further results are available, in which case nextSyncToken is provided.

kind

Type of the collection ("calendar#events").

default_reminders

The default reminders on the calendar for the authenticated user. These reminders apply to all events on this calendar that do not explicitly override them (i.e. do not have reminders.useDefault set to True).

description

Description of the calendar. Read-only.

items

List of events on the calendar.

updated

Last modification time of the calendar (as a RFC3339 timestamp). Read-only.

summary

Title of the calendar. Read-only.

etag

ETag of the collection.

next_sync_token

Token used at a later point in time to retrieve only the entries that have changed since this result was returned. Omitted if further results are available, in which case nextPageToken is provided.

time_zone

The time zone of the calendar. Read-only.

access_role

The user's access role for this calendar. Read-only. Possible values are: - "none" - The user has no access. - "freeBusyReader" - The user has read access to free/busy information. - "reader" - The user has read access to the calendar. Private events will appear to users with reader access, but event details will be hidden. - "writer" - The user has read and write access to the calendar. Private events will appear to users with writer access, and event details will be visible. - "owner" - The user has ownership of the calendar. This role has all of the permissions of the writer role with the additional ability to see and manipulate ACLs.