Struct google_calendar3::CalendarListEntry

pub struct CalendarListEntry {
    pub notification_settings: Option<CalendarListEntryNotificationSettings>,
    pub description: Option<String>,
    pub deleted: Option<bool>,
    pub primary: Option<bool>,
    pub access_role: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub foreground_color: Option<String>,
    pub default_reminders: Option<Vec<EventReminder>>,
    pub color_id: Option<String>,
    pub selected: Option<bool>,
    pub summary: Option<String>,
    pub etag: Option<String>,
    pub location: Option<String>,
    pub background_color: Option<String>,
    pub summary_override: Option<String>,
    pub time_zone: Option<String>,
    pub hidden: Option<bool>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get calendar list (response)

update calendar list (request|response)

patch calendar list (request|response)

insert calendar list (request|response)

Fields

notification_settings

The notifications that the authenticated user is receiving for this calendar.

description

Description of the calendar. Optional. Read-only.

deleted

Whether this calendar list entry has been deleted from the calendar list. Read-only. Optional. The default is False.

primary

Whether the calendar is the primary calendar of the authenticated user. Read-only. Optional. The default is False.

access_role

The effective access role that the authenticated user has on the calendar. Read-only. Possible values are: - "freeBusyReader" - Provides read access to free/busy information. - "reader" - Provides read access to the calendar. Private events will appear to users with reader access, but event details will be hidden. - "writer" - Provides read and write access to the calendar. Private events will appear to users with writer access, and event details will be visible. - "owner" - Provides ownership of the calendar. This role has all of the permissions of the writer role with the additional ability to see and manipulate ACLs.

id

Identifier of the calendar.

kind

Type of the resource ("calendar#calendarListEntry").

foreground_color

The foreground color of the calendar in the hexadecimal format "#ffffff". This property supersedes the index-based colorId property. To set or change this property, you need to specify colorRgbFormat=true in the parameters of the insert, update and patch methods. Optional.

default_reminders

The default reminders that the authenticated user has for this calendar.

color_id

The color of the calendar. This is an ID referring to an entry in the calendar section of the colors definition (see the colors endpoint). This property is superseded by the backgroundColor and foregroundColor properties and can be ignored when using these properties. Optional.

selected

Whether the calendar content shows up in the calendar UI. Optional. The default is False.

summary

Title of the calendar. Read-only.

etag

ETag of the resource.

location

Geographic location of the calendar as free-form text. Optional. Read-only.

background_color

The main color of the calendar in the hexadecimal format "#0088aa". This property supersedes the index-based colorId property. To set or change this property, you need to specify colorRgbFormat=true in the parameters of the insert, update and patch methods. Optional.

summary_override

The summary that the authenticated user has set for this calendar. Optional.

time_zone

The time zone of the calendar. Optional. Read-only.

hidden

Whether the calendar has been hidden from the list. Optional. The default is False.