Struct google_calendar3::Event

pub struct Event {
    pub creator: Option<EventCreator>,
    pub organizer: Option<EventOrganizer>,
    pub id: Option<String>,
    pub hangout_link: Option<String>,
    pub guests_can_invite_others: Option<bool>,
    pub attendees: Option<Vec<EventAttendee>>,
    pub start: Option<EventDateTime>,
    pub html_link: Option<String>,
    pub recurrence: Option<Vec<String>>,
    pub source: Option<EventSource>,
    pub etag: Option<String>,
    pub location: Option<String>,
    pub recurring_event_id: Option<String>,
    pub original_start_time: Option<EventDateTime>,
    pub status: Option<String>,
    pub updated: Option<String>,
    pub description: Option<String>,
    pub i_cal_uid: Option<String>,
    pub gadget: Option<EventGadget>,
    pub end_time_unspecified: Option<bool>,
    pub sequence: Option<i32>,
    pub visibility: Option<String>,
    pub guests_can_modify: Option<bool>,
    pub end: Option<EventDateTime>,
    pub attendees_omitted: Option<bool>,
    pub kind: Option<String>,
    pub locked: Option<bool>,
    pub created: Option<String>,
    pub color_id: Option<String>,
    pub anyone_can_add_self: Option<bool>,
    pub reminders: Option<EventReminders>,
    pub guests_can_see_other_guests: Option<bool>,
    pub summary: Option<String>,
    pub extended_properties: Option<EventExtendedProperties>,
    pub transparency: Option<String>,
    pub private_copy: Option<bool>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete events (none)

insert events (request|response)

instances events (none)

quick add events (response)

patch events (request|response)

import events (request|response)

move events (response)

update events (request|response)

watch events (none)

get events (response)

list events (none)

Fields

creator

The creator of the event. Read-only.

organizer

The organizer of the event. If the organizer is also an attendee, this is indicated with a separate entry in attendees with the organizer field set to True. To change the organizer, use the move operation. Read-only, except when importing an event.

id

Identifier of the event. When creating new single or recurring events, you can specify their IDs. Provided IDs must follow these rules: - characters allowed in the ID are those used in base32hex encoding, i.e. lowercase letters a-v and digits 0-9, see section 3.1.2 in RFC2938 - the length of the ID must be between 5 and 1024 characters - the ID must be unique per calendar Due to the globally distributed nature of the system, we cannot guarantee that ID collisions will be detected at event creation time. To minimize the risk of collisions we recommend using an established UUID algorithm such as one described in RFC4122.

hangout_link

An absolute link to the Google+ hangout associated with this event. Read-only.

guests_can_invite_others

Whether attendees other than the organizer can invite others to the event. Optional. The default is True.

attendees

The attendees of the event.

start

The (inclusive) start time of the event. For a recurring event, this is the start time of the first instance.

html_link

An absolute link to this event in the Google Calendar Web UI. Read-only.

recurrence

List of RRULE, EXRULE, RDATE and EXDATE lines for a recurring event. This field is omitted for single events or instances of recurring events.

source

Source of an event from which it was created; for example a web page, an email message or any document identifiable by an URL using HTTP/HTTPS protocol. Accessible only by the creator of the event.

etag

ETag of the resource.

location

Geographic location of the event as free-form text. Optional.

recurring_event_id

For an instance of a recurring event, this is the event ID of the recurring event itself. Immutable.

original_start_time

For an instance of a recurring event, this is the time at which this event would start according to the recurrence data in the recurring event identified by recurringEventId. Immutable.

status

Status of the event. Optional. Possible values are: - "confirmed" - The event is confirmed. This is the default status. - "tentative" - The event is tentatively confirmed. - "cancelled" - The event is cancelled.

updated

Last modification time of the event (as a RFC 3339 timestamp). Read-only.

description

Description of the event. Optional.

i_cal_uid

Event ID in the iCalendar format.

gadget

A gadget that extends this event.

end_time_unspecified

Whether the end time is actually unspecified. An end time is still provided for compatibility reasons, even if this attribute is set to True. The default is False.

sequence

Sequence number as per iCalendar.

visibility

Visibility of the event. Optional. Possible values are: - "default" - Uses the default visibility for events on the calendar. This is the default value. - "public" - The event is public and event details are visible to all readers of the calendar. - "private" - The event is private and only event attendees may view event details. - "confidential" - The event is private. This value is provided for compatibility reasons.

guests_can_modify

Whether attendees other than the organizer can modify the event. Optional. The default is False.

end

The (exclusive) end time of the event. For a recurring event, this is the end time of the first instance.

attendees_omitted

Whether attendees may have been omitted from the event's representation. When retrieving an event, this may be due to a restriction specified by the maxAttendee query parameter. When updating an event, this can be used to only update the participant's response. Optional. The default is False.

kind

Type of the resource ("calendar#event").

locked

Whether this is a locked event copy where no changes can be made to the main event fields "summary", "description", "location", "start", "end" or "recurrence". The default is False. Read-Only.

created

Creation time of the event (as a RFC 3339 timestamp). Read-only.

color_id

The color of the event. This is an ID referring to an entry in the event section of the colors definition (see the colors endpoint). Optional.

anyone_can_add_self

Whether anyone can invite themselves to the event. Optional. The default is False.

reminders

Information about the event's reminders for the authenticated user.

guests_can_see_other_guests

Whether attendees other than the organizer can see who the event's attendees are. Optional. The default is True.

summary

Title of the event.

extended_properties

Extended properties of the event.

transparency

Whether the event blocks time on the calendar. Optional. Possible values are: - "opaque" - The event blocks time on the calendar. This is the default value. - "transparent" - The event does not block time on the calendar.

private_copy

Whether this is a private event copy where changes are not shared with other copies on other calendars. Optional. Immutable. The default is False.